created super admin in database, change in user, doctor and medicalrecord

+++ PolarCape/HospitalManagementSystem/HospitalManagementSystem.Services/Implementation/MedicalRecordService.cs
@@ -1,6 +1,8 @@
 ﻿using HospitalManagementSystem.DataAccess.Interfaces;
+using HospitalManagementSystem.Domain.Models;
 using HospitalManagementSystem.DTO.MedicalRecordDtos;
 using HospitalManagementSystem.Services.Interfaces;
+using HospitalManagementSystem.Shared;
 using HospitaManagmentSystem.Mapper;
 
 namespace HospitalManagementSystem.Services.Implementation
@@ -17,15 +19,24 @@ namespace HospitalManagementSystem.Services.Implementation
             _petientsRepository = petientsRepository;
         }
 
-        public List<AllMedicalRecordByPatientId> AllMedicalRecordByPatientIds(int patientId)
+        public List<MedicalRecordDto> AllMedicalRecordByPatientIds(int patientId)
         {
            var patient = _petientsRepository.GetById(patientId);
-            return _medicalRecordRepository.GetAll().Where(x=>x.PatientId==patientId).Select(x=>x.ToAllMedicalRecordByPatientId()).ToList();
+            if(patient == null)
+            {
+                throw new PatientNotFoundException($"Patient with id: {patientId} not found"); 
+            }
+            var medicalRecords = _medicalRecordRepository.GetAll().Where(x => x.PatientId == patientId);
+            return medicalRecords.Select(x=>x.ToMedicalRecord()).ToList();
         }
 
-        public void CreateMedicalRecordForPatient(MedicalRecordDto medicalRecordDto, int userId)
+        public void CreateMedicalRecordForPatient(CreateMedicalRecordDto medicalRecordDto, int userId)
         {
             var doctor = _doctorRepository.GetAll().FirstOrDefault(x=>x.UserId == userId);
+            if(medicalRecordDto.PatientId == null)
+            {
+                throw new PatientNotFoundException($"Patient with id: {medicalRecordDto.PatientId} not found");
+            }
             var medicalReocord = medicalRecordDto.ToMedicalRecord();
             medicalReocord.DoctorId = doctor.Id;
             _medicalRecordRepository.Add(medicalReocord);
@@ -35,11 +46,56 @@ namespace HospitalManagementSystem.Services.Implementation
         {
             var doctor = _doctorRepository.GetAll().FirstOrDefault(x=>x.UserId==userId);
             var medicalRecord = _medicalRecordRepository.GetById(id);
-            if(medicalRecord.DoctorId != doctor.Id)
+            if (medicalRecord.DoctorId != doctor.Id)
             {
-                throw new Exception("Wrong entry");
+                throw new UnauthorizedAccessException("You do not have permission to delete this medical record");
             }
+
+            if (medicalRecord == null)
+            {
+                throw new MedicalRecordNotFoundException($"Medical record with id {id} does not exist");
+            }   
             _medicalRecordRepository.Delete(medicalRecord);
+        }
+
+        public List<MedicalRecordDto> GetAllMedicalRecord()
+        {
+            var medicalRecords = _medicalRecordRepository.GetAll();
+            return medicalRecords.Select(x=>x.ToMedicalRecord()).ToList();
+        }
+
+        public void UpdateMedicalRecord(UpdateMedicalRecord updateMedicalRecord, int userId)
+        {
+            var doctor = _doctorRepository.GetAll().FirstOrDefault(x=>x.UserId == userId);
+            var medicalRecord = _medicalRecordRepository.GetById(updateMedicalRecord.Id);
+
+            if (medicalRecord.DoctorId != doctor.Id)
+            {
+                throw new UnauthorizedAccessException("You do not have permission to delete this medical record");
+            }
+
+            if (medicalRecord == null)
+            {
+                throw new MedicalRecordNotFoundException($"Medical record with id {medicalRecord.Id} does not exist");
+            }
+            if (medicalRecord.StartDate.Date < DateTime.Now.Date)
+            {
+                throw new InvalidDataException("err stardate");
+            }
+            if (medicalRecord.EndDate < medicalRecord.StartDate)
+            {
+                throw new InvalidDataException("err enddate");
+            }
+
+
+            medicalRecord.Description = updateMedicalRecord.Description;
+            medicalRecord.StartDate = updateMedicalRecord.StartDate;
+            medicalRecord.EndDate = updateMedicalRecord.EndDate;
+            medicalRecord.Treatment = updateMedicalRecord.Description;
+            medicalRecord.Diagnosis = updateMedicalRecord.Diagnosis;
+            medicalRecord.DoctorId = doctor.Id;
+       
+            _medicalRecordRepository.Update(medicalRecord);
         }
     }
 }
